backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
firebase-admin==6.4.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
